+++ examples/PixelPusherServer/OctoWS2811Receiver.h
@@ -44,10 +44,9 @@ class OctoWS2811Receiver : public Receiver {
   }
 
   uint8_t stripFlags(int stripNum) const override;
-  void handleCommand(int command, const unsigned char *data, int len) override;
+  void handleCommand(int command, const uint8_t *data, int len) override;
   void startPixels(bool complete) override {}
-  void pixels(int stripNum, const unsigned char *pixels,
-              int pixelsPerStrip) override;
+  void pixels(int stripNum, const uint8_t *pixels, int pixelsPerStrip) override;
   void endPixels() override;
   void loop() override {}