+++ examples/PixelPusherServer/Receiver.h
@@ -39,8 +39,7 @@ class Receiver {
   virtual uint8_t stripFlags(int stripNum) const = 0;
 
   // Handles a PixelPusher command.
-  virtual void handleCommand(int command,
-                             const unsigned char *data, int len) {}
+  virtual void handleCommand(int command, const uint8_t *data, int len) {}
 
   // Starts receiving pixels from a packet. The `complete` parameter
   // indicates whether the complete set of pixels will be included in
@@ -48,7 +47,7 @@ class Receiver {
   virtual void startPixels(bool complete) = 0;
 
   // Processes pixels for one strip.
-  virtual void pixels(int stripNum, const unsigned char *pixels,
+  virtual void pixels(int stripNum, const uint8_t *pixels,
                       int pixelsPerStrip) = 0;
 
   // All the pixels from the packet have been sent to this receiver.